pyproject.toml
[tool.poetry]
name = "meetflow-ai"
version = "0.1.0"
description = "MeetFlow AI Multi-Agent System for Autonomous Enterprise Workflows"
authors = ["MeetFlow Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
alembic = "^1.13.1"
psycopg = {extras = ["binary"], version = "^3.1.17"}
pydantic = "^2.5.3"
google-generativeai = "^0.3.2"
slack-bolt = "^1.18.1"
httpx = "^0.26.0"
langgraph = "^0.0.26"
redis = "^5.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
